Fix int32 overflow in SHT40 temperature conversion
175000 * raw_temp overflows int32_t when raw_temp > ~12270 (which is
any temperature above -11C). Changed to int64_t multiplication in
all 6 files that do the conversion.

Before: Temperature: -43.-19 C (wrong — int32 overflow)
After:  Temperature: 22.29 C   (correct)

+++ app/src/05_soft_reset.c
@@ -113,8 +113,8 @@ static int sht40_measure(const struct device *i2c_dev)
 	uint16_t raw_temp = ((uint16_t)buf[0] << 8) | buf[1];
 	uint16_t raw_hum  = ((uint16_t)buf[3] << 8) | buf[4];
 
-	int temp_milli = -45000 + (175000 * (int32_t)raw_temp) / 65535;
-	int hum_milli  = -6000  + (125000 * (int32_t)raw_hum)  / 65535;
+	int temp_milli = -45000 + (int)((175000LL * (int64_t)raw_temp) / 65535);
+	int hum_milli  = -6000  + (int)((125000LL * (int64_t)raw_hum)  / 65535);
 
 	if (hum_milli < 0) hum_milli = 0;
 	if (hum_milli > 100000) hum_milli = 100000;
